+++ Sources/VComponents/Components/Buttons/Stretched Button (Loading)/VLoadingStretchedButton.swift
@@ -101,6 +101,7 @@ public struct VLoadingStretchedButton<Label>: View where Label: View {
                 let internalState: VLoadingStretchedButtonInternalState = internalState(baseButtonState)
                 
                 buttonLabel(internalState: internalState)
+                    .contentShape(Rectangle()) // Registers gestures even when clear
                     .frame(height: uiModel.height)
                     .cornerRadius(uiModel.cornerRadius) // Prevents large content from overflowing
                     .background(background(internalState: internalState)) // Has own rounding